feat: improve cache on contract, tx and block page
