+++ packages/app/next.config.mjs
@@ -39,6 +39,30 @@ const config = {
         destination: '/ui/index.html',
         permanent: false,
       },
+      // Redirect to main page
+      {
+        source: '/tx/:id',
+        destination: '/tx/:id/simple',
+        permanent: false,
+      },
+      // Redirect to main page
+      {
+        source: '/block/:id',
+        destination: '/block/:id/simple',
+        permanent: false,
+      },
+      // Redirect to main page
+      {
+        source: '/contract/:id',
+        destination: '/contract/:id/assets',
+        permanent: false,
+      },
+      // Redirect to main page
+      {
+        source: '/',
+        destination: '/transactions?page=1',
+        permanent: false,
+      },
     ];
   },
   webpack: (config) => {
