+++ packages/app/src/app/api/graphql/route.ts
@@ -13,9 +13,6 @@ const executor = createExecutor(async ({ body }) => {
     headers: {
       'Content-Type': 'application/json',
     },
-    next: {
-      revalidate: Infinity,
-    },
   }).then((res) => {
     console.timeEnd('request');
     return res.json();
@@ -35,14 +32,6 @@ const { handleRequest: _handleRequest } = createYoga({
 });
 
 const handleRequest = _handleRequest;
-// // Use to delay more the application
-// // eslint-disable-next-line @typescript-eslint/no-explicit-any
-// const handleRequest = async (...args: any[]) => {
-//   // console.log(args);
-//   const response = _handleRequest(...args);
-//   await setTimeout(5000);
-//   return response;
-// };
 
 export {
   handleRequest as GET,
