+++ packages/app/src/app/block/[id]/layout.tsx
@@ -20,4 +20,5 @@ export default function BlockLayout({
   );
 }
 
+export const fetchCache = 'force-static';
 export const invalidate = Infinity;
