--- packages/app/src/app/block/[id]/page.tsx
@@ -1,9 +1,0 @@
-import { redirect } from 'next/navigation';
-
-export default async function main({
-  params: { id },
-}: {
-  params: { id: string };
-}) {
-  redirect(`/block/${id}/simple`);
-}
