+++ packages/app/src/app/contract/[address]/assets/page.tsx
@@ -1,0 +1,23 @@
+import { Suspense } from 'react';
+import { AssetsSkeleton } from '~/systems/Asset/components/AssetsSkeleton';
+import { ContractAssets } from '~/systems/Contract/screens/ContractAsset';
+
+type ContractProps = {
+  params: {
+    address: string;
+    tab: string;
+  };
+};
+
+export default async function ContractPage({
+  params: { address },
+}: ContractProps) {
+  return (
+    <Suspense fallback={<AssetsSkeleton />}>
+      <ContractAssets id={address} />
+    </Suspense>
+  );
+}
+
+export const fetchCache = 'force-dynamic';
+export const revalidate = 10;
