+++ packages/app/src/app/contract/[address]/layout.tsx
@@ -9,3 +9,6 @@ export default function Layout({
 }) {
   return <ContractLayout id={address}>{children}</ContractLayout>;
 }
+
+export const fetchCache = 'force-static';
+export const revalidate = Infinity;
