--- packages/app/src/app/contract/[address]/loading.tsx
@@ -1,5 +1,0 @@
-import { AssetsSkeleton } from '~/systems/Asset/components/AssetsSkeleton';
-
-export default function Loading() {
-  return <AssetsSkeleton />;
-}
