--- packages/app/src/app/contract/[address]/page.tsx
@@ -1,9 +1,0 @@
-import { redirect } from 'next/navigation';
-
-export default async function main({
-  params: { address },
-}: {
-  params: { address: string };
-}) {
-  redirect(`/contract/${address}/assets`);
-}
