--- packages/app/src/app/tx/[id]/page.tsx
@@ -1,11 +1,0 @@
-import { redirect } from 'next/navigation';
-
-export default async function main({
-  params: { id },
-}: {
-  params: { id: string };
-}) {
-  redirect(`/tx/${id}/simple`);
-}
-
-export const revalidate = Infinity;
